Move cbloq_to_cirq logic into cirq_conversion module
